+++ integration_tests/elemental_05.py
@@ -1,5 +1,5 @@
 from lpython import i32, f64, f32
-from numpy import empty, sinh, cosh, reshape, int32, float64, sin
+from numpy import empty, sinh, cosh, reshape, int32, float32, float64, sin
 
 def verify1d(array: f32[:], result: f32[:], size: i32):
     i: i32
@@ -24,8 +24,8 @@ def verifynd(array: f64[:, :, :, :], result: f64[:, :, :, :], size1: i32, size2:
 def elemental_sinh():
     i: i32; j: i32; k: i32; l: i32; size: i32;
 
-    array1d: f32[10] = empty(10)
-    sinh1d: f32[10] = empty(10)
+    array1d: f32[10] = empty(10, dtype=float32)
+    sinh1d: f32[10] = empty(10, dtype=float32)
 
     for i in range(10):
         array1d[i] = f32(f64(i)/10.0)
@@ -33,8 +33,8 @@ def elemental_sinh():
     sinh1d = sinh(sinh(array1d))
     verify1d(array1d, sinh1d, 10)
 
-    arraynd: f64[40, 10, 16, 2] = empty((40, 10, 16, 2))
-    sinhnd: f64[40, 10, 16, 2] = empty((40, 10, 16, 2))
+    arraynd: f64[40, 10, 16, 2] = empty((40, 10, 16, 2), dtype=float64)
+    sinhnd: f64[40, 10, 16, 2] = empty((40, 10, 16, 2), dtype=float64)
     size = 40 * 10 * 16 * 2
 
     for i in range(40):
@@ -58,12 +58,12 @@ def verify2d(array: f64[:, :], result: f64[:, :], size1: i32, size2: i32):
 def elemental_cosh():
     i: i32; j: i32
 
-    array2d: f64[20, 10] = empty((20, 10))
-    cosh2d: f64[20, 10] = empty((20, 10))
+    array2d: f64[20, 10] = empty((20, 10), dtype=float64)
+    cosh2d: f64[20, 10] = empty((20, 10), dtype=float64)
 
     for i in range(20):
         for j in range(10):
-                array2d[i, j] = float(i + 2*j)/200.0  
+                array2d[i, j] = float(i + 2*j)/200.0
 
     cosh2d = cosh(5.0 + (array2d))**2.0
     verify2d(array2d, cosh2d, 20, 10)
@@ -72,8 +72,8 @@ def elemental_cosh_():
     i: i32
     j: i32
 
-    array2d: f64[20, 10] = empty((20, 10))
-    cosh2d: f64[20, 10] = empty((20, 10))
+    array2d: f64[20, 10] = empty((20, 10), dtype=float64)
+    cosh2d: f64[20, 10] = empty((20, 10), dtype=float64)
 
     for i in range(20):
         for j in range(10):
@@ -111,7 +111,7 @@ def elemental_trig_identity():
                   cosh(arraynd/4.0) * cosh(arraynd/2.0) -
                   sinh(arraynd/4.0) * sinh(arraynd/2.0))
 
-    newshape: i32[1] = empty(1, dtype=int)
+    newshape: i32[1] = empty(1, dtype=int32)
     newshape[0] = 400
 
     observed1d_1 = reshape(identity1, newshape)
