+++ integration_tests/elemental_12.py
@@ -1,5 +1,5 @@
 from lpython import i32, f32, f64
-from numpy import empty, floor, ceil, sqrt, reshape
+from numpy import empty, floor, ceil, sqrt, reshape, int32, float32, float64
 
 def elemental_floor64():
     i: i32
@@ -9,9 +9,9 @@ def elemental_floor64():
     eps: f32
     eps = f32(1e-6)
 
-    arraynd: f64[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    arraynd: f64[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float64)
 
-    newshape: i32[1] = empty(1, dtype=int)
+    newshape: i32[1] = empty(1, dtype=int32)
     newshape[0] = 16384
 
     for i in range(32):
@@ -20,13 +20,13 @@ def elemental_floor64():
                 for l in range(4):
                     arraynd[i, j, k, l] = f64((-1)**l) * sqrt(float(i + j + k + l))
 
-    observed: f64[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    observed: f64[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float64)
     observed = floor(arraynd)
 
-    observed1d: f64[16384] = empty(16384)
+    observed1d: f64[16384] = empty(16384, dtype=float64)
     observed1d = reshape(observed, newshape)
 
-    array: f64[16384] = empty(16384)
+    array: f64[16384] = empty(16384, dtype=float64)
     array = reshape(arraynd, newshape)
 
     for i in range(16384):
@@ -41,7 +41,7 @@ def elemental_floor32():
     eps: f32
     eps = f32(1e-6)
 
-    arraynd: f32[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    arraynd: f32[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float32)
 
     for i in range(32):
         for j in range(16):
@@ -49,7 +49,7 @@ def elemental_floor32():
                 for l in range(4):
                     arraynd[i, j, k, l] = f32(f64((-1)**l) * sqrt(float(i + j + k + l)))
 
-    observed: f32[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    observed: f32[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float32)
     observed = floor(arraynd)
 
     for i in range(32):
@@ -67,9 +67,9 @@ def elemental_ceil64():
     eps: f32
     eps = f32(1e-6)
 
-    arraynd: f64[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    arraynd: f64[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float64)
 
-    newshape: i32[1] = empty(1, dtype=int)
+    newshape: i32[1] = empty(1, dtype=int32)
     newshape[0] = 16384
 
     for i in range(32):
@@ -78,13 +78,13 @@ def elemental_ceil64():
                 for l in range(4):
                     arraynd[i, j, k, l] = f64((-1)**l) * sqrt(float(i + j + k + l))
 
-    observed: f64[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    observed: f64[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float64)
     observed = ceil(arraynd)
 
-    observed1d: f64[16384] = empty(16384)
+    observed1d: f64[16384] = empty(16384, dtype=float64)
     observed1d = reshape(observed, newshape)
 
-    array: f64[16384] = empty(16384)
+    array: f64[16384] = empty(16384, dtype=float64)
     array = reshape(arraynd, newshape)
 
     for i in range(16384):
@@ -99,7 +99,7 @@ def elemental_ceil32():
     eps: f32
     eps = f32(1e-6)
 
-    arraynd: f32[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    arraynd: f32[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float32)
 
     for i in range(32):
         for j in range(16):
@@ -107,7 +107,7 @@ def elemental_ceil32():
                 for l in range(4):
                     arraynd[i, j, k, l] = f32(f64((-1)**l) * sqrt(float(i + j + k + l)))
 
-    observed: f32[32, 16, 8, 4] = empty((32, 16, 8, 4))
+    observed: f32[32, 16, 8, 4] = empty((32, 16, 8, 4), dtype=float32)
     observed = ceil(arraynd)
 
     for i in range(32):
